fix test mock
